refactor: move API route telemetry middleware
